fix : deleted class PlayersLog and moved its content as private method in FileManipulation class, added line in Fixture
