+++ lib/file_manipulation.rb
@@ -1,4 +1,3 @@
-require_relative 'players_log'
 require 'json'
 
 class FileManipulation
@@ -13,8 +12,7 @@ class FileManipulation
   end
 
   def info_log_to_json
-    players = PlayersLog.new(@path_to_file)
-    hash = { @path_to_file => { 'lines' => number_of_lines, 'players' => players.players_names } }
+    hash = { @path_to_file => { 'lines' => number_of_lines, 'players' => players_names } }
     hash.to_json
   end
 
@@ -22,5 +20,17 @@ class FileManipulation
 
   def number_of_lines
     File.readlines(@path_to_file).size
+  end
+
+  def players_names
+    all_players = []
+    File.readlines(@path_to_file).each do |line|
+      if line.include?('ClientUserinfoChanged:')
+        splited_line_with_name_in_position_one = line.split(/[\\\\]/)
+        player_name = splited_line_with_name_in_position_one[1]
+        all_players.push(player_name) unless all_players.include?(player_name)
+      end
+    end
+    all_players
   end
 end
